[DHBLOG] 백엔드 댓글 삭제 시 작성자 일치 여부

+++ src/main/java/com/dhkim/blog/post/service/PostServiceImpl.java
@@ -243,7 +243,17 @@ public class PostServiceImpl implements PostService{
     @Override
     public String replyDelete(HttpServletRequest request, String id) {
         try{
-            replyRepository.deleteById(id);
+            Optional<Reply> optionalReply = replyRepository.findById(id);
+            if (optionalReply.isPresent()) {
+                Reply reply = optionalReply.get();
+                HttpSession session = request.getSession();
+
+                if(reply.getNickname().equals(session.getAttribute("userNickname").toString())){
+                    replyRepository.deleteById(id);
+                }else{
+                    return "error";
+                }
+            }
         }catch (Exception e){
             return "error";
         }
